Fixed a nasty bug that took me ages to track down. The calculate function was writing way out of bounds on one of it's pub arrays. This didn't cause a seg fault as all the memory around that array belongs to this application. It was tricky to track down as because it only showed up when you called adjacent overwritten function pointers.
Other than asking the user to be more careful (I'd defined an array much shorter than it should have been), the other option is to introduce bounds checking in the macros, maybe mixing in debug macros so you can switch off these checks when optimising. This is doable and will be considered.