Backed out previous change. The reason I didn't think of this before is that having objects on the same tier subscribe to one another - even if they don't read from any data that's currently being written to - buggers up the timestamps.
This was introduced to enable a homogeneous tier full of objects to run concurrently _and_ subscribe to one another. The problem of how concurrently running objects can subscribe to one another is solved by all objects storing a historical record of previous published data and object A reading object B's published data at t-1.

This is required to allow trad4 to model n-body problems of various sorts (orrery is one, but I've long had plans to write an app called perfect_gas, and use it to try and measure the speed of sound analytically at various pressures).

The problem is the T1 objects get run asynchronously relative to one another, so the need_refresh mechanism breaks down.

So the much simpler solution for orrery would be to just introduce a null object at T2 to which all the T1 objects subscribe, forcing all the T1 objects to update each cycle (which is what we want for this kind of problem). Fortunately we already have such an (non-null) object at T2 - monitor. Update to orrery on it's way. Thorough testing to follow...

I know this is a long checkin comment but hey - it's been a while.