Ok here's the deal.
The checkin below (r1306) was confused and misguided. There was a problem with getting "a homogeneous tier full of objects to run concurrently _and_ subscribe to one an other", but it was caused by the structure of orrery, not r1273 which r1306 was backing out.

Checkin r1273 does have desirable functionality, but also caused an unrelated performance problem which was how I got so confused.

So, this checkin re-applies r1273 and fixes the performance problem it caused. A fix to orrery to solve the timing problem will be made shortly.